make order card

+++ client/src/Pages/Home/HomeComponents/AllUsers/UsersCard/UsersCard.jsx
@@ -22,11 +22,18 @@ const UsersCard = ({user,loading,reFetch}) => {
   }
 
   const userDeleteHandler = async (id) => {
+    const isConfirmed = window.confirm(
+      `আপনি কি নিশ্চিত যে ${name} নামের সদস্য কে মুছে ফেলতে চান?`
+  );
+  if (!isConfirmed) {
+      toast("সদস্য কে মুছে ফেলার কাজ বাতিল করা হয়েছে।", { icon: "⚠️" });
+      return;
+  }
     try{
       const deleteUserResult = await deleteUser(id);
       if(deleteUserResult?.deletedCount){
         reFetch();
-        toast.success("সফল ভাবে ইউজারকে মুছে ফেলা হয়েছে")
+        toast.success("সফল ভাবে সদস্য কে মুছে ফেলা হয়েছে")
       }
     }
     catch(err){
@@ -36,7 +43,7 @@ const UsersCard = ({user,loading,reFetch}) => {
 
   if(loading) return <Loader />
     return <div className="w-[95%] md:w-[600px] shadow-lg">
-    <div className="flex border border-primary items-center justify-between p-3 ">
+    <div className="flex border border-primary items-center justify-between p-[14px] ">
         <div className="font-poppins">
             <h3 className="text-lg font-semibold">{name}</h3>
             <p className="text-sm text-gray-600">{email}</p>
@@ -44,10 +51,10 @@ const UsersCard = ({user,loading,reFetch}) => {
         </div>
 
         <div className="grid grid-cols-1">
-            <button onClick={()=>makeAdminHandler(_id)} className="bg-primary py-[2px] text-sm text-neutral shadow-lg px-2 hover:bg-[#ff1c68] transition-all capitalize rounded">
+            <button onClick={()=>makeAdminHandler(_id)} className="bg-primary text-sm text-neutral shadow-lg p-[6px] hover:bg-[#ff1c68] transition-all capitalize rounded">
               এডমিন করুন
             </button>
-            <button onClick={()=>userDeleteHandler(_id)} className="bg-primary py-[2px] text-sm text-neutral shadow-lg px-2 mt-4 hover:bg-[#ff1c68] transition-all capitalize rounded">
+            <button onClick={()=>userDeleteHandler(_id)} className="bg-primary  text-sm text-neutral shadow-lg p-[6px] mt-4 hover:bg-[#ff1c68] transition-all capitalize rounded">
               মুছে ফেলুন
             </button>
         </div>
